Avoid a notice
If there are no categories and tags the line starts with a pipe, which is not
very nice. Maybe say that there was no category or so?

I think it would be better to generate all the parts seperately and then implode
the seperators in. However this would need get_comments_popup_link which is not
currently available: https://core.trac.wordpress.org/ticket/17763